fram: Make advanced board-specific methods optional
Make all but the most basic board-specific methods optional.
Optional methods that are not available will fall through to the
"unknown command" case.

Make sure that all commands that are executed cause termination of the
script, using either exec (for external commands), or exit (for methods
that can be shell functions).

Signed-off-by: Geert Uytterhoeven <geert@linux-m68k.org>