fram: Call Wake-on-LAN through board-specific method
If /etc/ethers is shared with e.g. dnsmasq, it may contain MAC addresses
for Ethernet interfaces that do not support Wake-on-LAN.  Moreover, some
systems may use a different external command than "wakeonlan", which may
not rely on /etc/ethers at all.

Hence replace the check in /etc/ethers and the direct call to the
wakeonlan command by a check for a board-specific "wol-$BOARD" method,
and a call into that method.

Signed-off-by: Geert Uytterhoeven <geert@linux-m68k.org>